Made the camera into a continuous video + bug fix
The bug fix was the recogStart function returning blank picture, if a face was detected and then would delete the window. And i made the script run faster by using hog instead of cnn

The continuous video only happens when a face is recognised, if a face isnt recognised, the function ends up recursing itself so fast, it crashes and dies. Need to fix this asap